WIP: Add condition to catch writing zero bytes to toc file
This catch was inspired by a strange issue -- one user of gpbackup
happened to take a backup which properly contained data, but had an
empty toc file. The file was properly created, but no bytes were written
to it.

This allowed the backup to exit successfully, however rows were lost on
restore, as there were no byte offsets to reference in the segments toc.
This occued on two(out of 200+) segments

**NOTE:** Remove the commented out section prior to merging. That is
simply there for manual validation of this error.

Authored-by: Trevor Yacovone <tyacovone@pivotal.io>